Added CLEAN_BLOG_CONTENT_STYLE setting
The `CLEAN_BLOG_CONTENT_STYLE` allows the main page content to be
centered with padding on the sides.  This is similar _Clean Blog_:

  - https://startbootstrap.com/template-overviews/clean-blog/

This setting takes precidence over `HIDE_SIDEBAR`, `ABOUT_ME`, and
`AVATAR`, hiding the sidebar completely when `CLEAN_BLOG_CONTENT_STYLE`
is set to `True`.

Also cleaned up how the `class`es are assigned in the main content div.
Things were starting to get unreadable in there IMHO.